+++ sushii-2/src/commands/tags/info.rs
@@ -1,7 +1,7 @@
+use chrono::{DateTime, Utc};
 use serenity::framework::standard::{macros::command, Args, CommandResult};
 use serenity::model::prelude::*;
 use serenity::prelude::*;
-use chrono::{DateTime, Utc};
 
 use crate::model::sql::*;