+++ sushii-2/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -  Add more bot information to `stats` / `about` command.
+-  Display created datetime on `tag info`
+-  `feed delete` command
 
 ## [0.2.0] - 2021-02-18